worker/requirements.txt
redis==5.0.1
asyncpg==0.30.0
SQLAlchemy==2.0.43
httpx==0.27.0
python-dotenv==1.1.1
